Improve nested exectution
- Enables nested definition of node urls in the distribution options
- Randomly deep nesting supported
- If maxDepth is bigger than there are actual layers in the nested url
definition, the last level of urls is used for the rest of the layers.
This is not very efficient fo execution, but offers nice flexibility
for testing larger depths
- Also, if executing nodes have multiple cores, it’s simpler to run
nested executions than if avery layer would need it’s own url
definitions. In this way, nodes can distribute the load by themselves
internally